Extract gatsby-source-sylius
