+++ package.json
@@ -18,14 +18,12 @@
     "gatsby-plugin-manifest": "^2.2.27",
     "gatsby-plugin-offline": "^3.0.19",
     "gatsby-plugin-react-helmet": "^3.1.14",
-    "gatsby-plugin-remote-images": "^1.0.7",
     "gatsby-plugin-sharp": "^2.2.37",
     "gatsby-plugin-smoothscroll": "^1.0.4",
     "gatsby-plugin-styled-components": "^3.1.16",
     "gatsby-plugin-typography": "^2.3.20",
     "gatsby-source-filesystem": "^2.1.36",
     "gatsby-transformer-sharp": "^2.3.3",
-    "got": "^9.6.0",
     "intersection-observer": "^0.7.0",
     "lodash.get": "^4.4.2",
     "prop-types": "^15.7.2",
@@ -61,7 +59,8 @@
     "serve": "gatsby serve",
     "deploy": "gatsby build --prefix-paths && gh-pages -d public",
     "test": "echo \"Write tests! -> https://gatsby.dev/unit-testing\" && exit 1",
-    "clean": "gatsby clean"
+    "clean": "gatsby clean",
+    "postinstall": "cd plugins/gatsby-source-sylius/ && npm install"
   },
   "repository": {
     "type": "git",
