+++ plugins/gatsby-source-sylius/index.js
@@ -1,0 +1,1 @@
+// mandatory for gatsby
