+++ plugins/gatsby-source-sylius/validateOptions.js
@@ -1,0 +1,7 @@
+module.exports = ({ reporter }, options) => {
+  if (!options.url) {
+    reporter.panic(
+      "Problems with gatsby-source-sylius options: url is mandatory"
+    );
+  }
+};
