--- sylius-consumer/index.js
@@ -1,11 +1,0 @@
-// const makeTypes = require("./makeTypes");
-const createResolvers = require("./createResolvers");
-// const makeNodes = require("./makeNodes");
-// const queries = require("./queries");
-
-module.exports = {
-  // makeTypes,
-  createResolvers,
-  // makeNodes,
-  // queries,
-};
